Split up write_docs.py into multiple files.
