+++ src/antsibull_docs/cli/doc_commands/plugin.py
@@ -25,7 +25,7 @@ from ...docs_parsing import AnsibleCollectionMetadata
 from ...docs_parsing.fqcn import get_fqcn_parts, is_fqcn
 from ...jinja2.environment import doc_environment
 from ...utils.collection_name_transformer import CollectionNameTransformer
-from ...write_docs import write_plugin_rst
+from ...write_docs.plugins import write_plugin_rst
 from .stable import normalize_plugin_info
 
 mlog = log.fields(mod=__name__)
